feat: implement loading of event
